requirements.txt
pandas
plotly
dash
dash_core_components
dash_html_components
dash_bootstrap_components
requests
requests_oauthlib
python-twitter